requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
loguru==0.6.0
lxml==5.2.2
Nuitka==2.2.3
ordered-set==4.1.0
psutil==5.9.8
PySide6==6.7.0
PySide6_Addons==6.7.0
PySide6_Essentials==6.7.0
python-dateutil==2.9.0.post0
requests==2.31.0
shiboken6==6.7.0
six==1.16.0
soupsieve==2.5
urllib3==2.2.1
zstandard==0.22.0
